API Perform Travel and Attack

+++ pkg/common/types/travel.go
@@ -13,20 +13,20 @@ type TravelPostPayload struct {
 }
 
 type TravelPostResponse struct {
-	Text                    string `json:"text"`
-	ResultText              string `json:"resultText"`
-	RewardType              string `json:"rewardType"`
-	RewardAmount            int64  `json:"rewardAmount"`
-	CurrentEXP              int64  `json:"currentEXP"`
-	CurrentGold             int64  `json:"currentGold"`
-	Level                   int64  `json:"level"`
-	NextWait                int64  `json:"nextwait"`
-	IsNpc                   bool   `json:"is_npc"`
-	NpcName                 string `json:"npc_name"`
-	NpcAvatar               string `json:"npc_avatar"`
-	UserAmount              int64  `json:"userAmount"`
-	GuildUserAmount         int64  `json:"guild_userAmount"`
-	GuildDropRatePercentage int    `json:"guild_droprate_percentage"`
-	GuildGoldIncrease       int    `json:"guild_gold_increase"`
-	GuildExpIncrease        int    `json:"guild_exp_increase"`
+	Text                    string  `json:"text"`
+	ResultText              string  `json:"resultText"`
+	RewardType              string  `json:"rewardType"`
+	RewardAmount            float64 `json:"rewardAmount"`
+	CurrentEXP              int64   `json:"currentEXP"`
+	CurrentGold             int64   `json:"currentGold"`
+	Level                   int64   `json:"level"`
+	NextWait                int64   `json:"nextwait"`
+	IsNpc                   bool    `json:"is_npc"`
+	NpcName                 string  `json:"npc_name"`
+	NpcAvatar               string  `json:"npc_avatar"`
+	UserAmount              int64   `json:"userAmount"`
+	GuildUserAmount         int64   `json:"guild_userAmount"`
+	GuildDropRatePercentage float64 `json:"guild_droprate_percentage"`
+	GuildGoldIncrease       float64 `json:"guild_gold_increase"`
+	GuildExpIncrease        float64 `json:"guild_exp_increase"`
 }
